feat: 添加视频进度管理功能，集成 Bilibili API，支持定时任务调度和数据持久化。更新项目结构，新增相关文档，完善配置加载逻辑。
- 在 go.mod 中添加 cron 库依赖。
- 实现视频进度获取和记录的应用服务。
- 完善数据库模型，支持视频观看进度的存储。
- 新增调度器，定时获取视频进度并记录。
- 更新 README 文档，描述应用层和基础设施层的职责。

+++ internal/infrastructure/bilibili/client.go
@@ -10,17 +10,17 @@ import (
 
 const (
 	baseURL = "https://api.bilibili.com/x/player/wbi/v2"
-	// TODO: Dynamically load SESSDATA instead of hardcoding
+	// TODO: 动态加载 SESSDATA 而不是硬编码
 	sessData = "SESSDATA=a486b214%2C1761995605%2Ce8064%2A51CjAjbbi4oaDoEXCN7yjPThSija81Url7d8duiZqLF-IVReywvw-pC5bbiw4O0IFLqZwSVlZUZ3lDMktiU2NSYTMwYTVqRnoxMmoxRG5jVEVQaVlXb0tzdm1zX1k3VlpKcGQ2aFk2SzMyMVJ5SlNRc1g4MmtNckdkTFp0RjZJM0V0WnpyMnRUMUNBIIEC"
 )
 
-// VideoProgressData 定义 Bilibili 视频进度 API 响应中 data 字段的结构体
+// VideoProgressData 定义 Bilibili 视频进度 API 响应中 data 字段的结构体。
 type VideoProgressData struct {
 	Aid                     int64         `json:"aid"`
 	Bvid                    string        `json:"bvid"`
 	AllowBp                 bool          `json:"allow_bp"`
 	NoShare                 bool          `json:"no_share"`
-	Cid                     int64         `json:"cid"`
+	Cid                     int64         `json:"cid"` // 注意：API 返回的是当前视频页面的 CID，而不是上次播放的 CID
 	MaxLimit                int           `json:"max_limit"`
 	PageNo                  int           `json:"page_no"`
 	HasNext                 bool          `json:"has_next"`
@@ -35,8 +35,8 @@ type VideoProgressData struct {
 	AnswerStatus            int           `json:"answer_status"`
 	BlockTime               int           `json:"block_time"`
 	Role                    string        `json:"role"`
-	LastPlayTime            int           `json:"last_play_time"` // 观看进度，单位毫秒
-	LastPlayCid             int64         `json:"last_play_cid"`
+	LastPlayTime            int64         `json:"last_play_time"` // 观看进度，单位毫秒
+	LastPlayCid             int64         `json:"last_play_cid"`  // 上次播放的视频分 P ID
 	NowTime                 int64         `json:"now_time"`
 	OnlineCount             int           `json:"online_count"`
 	NeedLoginSubtitle       bool          `json:"need_login_subtitle"`
@@ -59,7 +59,7 @@ type VideoProgressData struct {
 	IsUpowerExclusiveWithQa bool          `json:"is_upower_exclusive_with_qa"`
 }
 
-// IpInfo IP 相关信息
+// IpInfo IP 相关信息。
 type IpInfo struct {
 	Ip       string `json:"ip"`
 	ZoneIp   string `json:"zone_ip"`
@@ -69,7 +69,7 @@ type IpInfo struct {
 	City     string `json:"city"`
 }
 
-// LevelInfo 等级信息
+// LevelInfo 等级信息。
 type LevelInfo struct {
 	CurrentLevel int   `json:"current_level"`
 	CurrentMin   int   `json:"current_min"`
@@ -78,7 +78,7 @@ type LevelInfo struct {
 	LevelUp      int64 `json:"level_up"`
 }
 
-// VipInfo VIP 相关信息
+// VipInfo VIP 相关信息。
 type VipInfo struct {
 	Type               int        `json:"type"`
 	Status             int        `json:"status"`
@@ -96,7 +96,7 @@ type VipInfo struct {
 	AvatarIcon         AvatarIcon `json:"avatar_icon"`
 }
 
-// VipLabel VIP 标签信息
+// VipLabel VIP 标签信息。
 type VipLabel struct {
 	Path                  string `json:"path"`
 	Text                  string `json:"text"`
@@ -112,18 +112,18 @@ type VipLabel struct {
 	ImgLabelUriHantStatic string `json:"img_label_uri_hant_static"`
 }
 
-// AvatarIcon 头像图标信息
+// AvatarIcon 头像图标信息。
 type AvatarIcon struct {
 	IconResource interface{} `json:"icon_resource"` // 结构未知，暂用 interface{}
 }
 
-// Options 播放器选项
+// Options 播放器选项。
 type Options struct {
 	Is360      bool `json:"is_360"`
 	WithoutVip bool `json:"without_vip"`
 }
 
-// OnlineSwitch 在线开关配置
+// OnlineSwitch 在线开关配置。
 type OnlineSwitch struct {
 	EnableGrayDashPlayback string `json:"enable_gray_dash_playback"`
 	NewBroadcast           string `json:"new_broadcast"`
@@ -131,18 +131,18 @@ type OnlineSwitch struct {
 	SubtitleSubmitSwitch   string `json:"subtitle_submit_switch"`
 }
 
-// Fawkes 配置信息
+// Fawkes 配置信息。
 type Fawkes struct {
 	ConfigVersion int `json:"config_version"`
 	FfVersion     int `json:"ff_version"`
 }
 
-// ShowSwitch 显示开关配置
+// ShowSwitch 显示开关配置。
 type ShowSwitch struct {
 	LongProgress bool `json:"long_progress"`
 }
 
-// ElecHighLevel 高能等级信息
+// ElecHighLevel 高能等级信息。
 type ElecHighLevel struct {
 	PrivilegeType int    `json:"privilege_type"`
 	Title         string `json:"title"`
@@ -156,8 +156,8 @@ type ElecHighLevel struct {
 	QaTitle       string `json:"qa_title"`
 }
 
-// VideoProgressResponse 定义 Bilibili 视频进度 API 的响应结构体
-// 注意：这里需要根据实际 API 返回的 JSON 结构进行定义
+// VideoProgressResponse 定义 Bilibili 视频进度 API 的响应结构体。
+// 注意：这里需要根据实际 API 返回的 JSON 结构进行定义。
 type VideoProgressResponse struct {
 	Code    int               `json:"code"`
 	Message string            `json:"message"`
@@ -165,14 +165,14 @@ type VideoProgressResponse struct {
 	Data    VideoProgressData `json:"data"` // 更新为具体的 Data 结构体
 }
 
-// Client Bilibili API 客户端结构体
+// Client Bilibili API 客户端结构体。
 type Client struct {
 	httpClient *http.Client
 	baseURL    string
 	sessData   string
 }
 
-// NewClient 创建一个新的 Bilibili API 客户端实例
+// NewClient 创建一个新的 Bilibili API 客户端实例。
 func NewClient() *Client {
 	return &Client{
 		httpClient: &http.Client{},
@@ -181,9 +181,9 @@ func NewClient() *Client {
 	}
 }
 
-// GetVideoProgress 调用 Bilibili API 获取指定视频的观看进度
+// GetVideoProgress 调用 Bilibili API 获取指定视频的观看进度。
 // aid: 视频的 AV 号 (不带 'av' 前缀)
-// cid: 视频的分 P ID
+// cid: 视频的分 P ID (当前页面的 CID)
 func (c *Client) GetVideoProgress(aid, cid string) (*VideoProgressResponse, error) {
 	// 构建请求 URL
 	requestURL, err := url.Parse(c.baseURL)
